Add support for '--auto arch' (#4)
* Make MinioServer.Config.minio_arch/0 public

I plan to use this for automatically detecting a MinIO architecture
which matches the current machine.

This won't do the right thing on Linux/ARM setups, but unfortunately I
cannot test what kind of architecture to use in those cases. Since they
never seemed to be supported very well, let's just ignore them for the
time being.

* Add recognition of '--arch auto' parameter

This extends the handling of the '--arch' argument such that it
recognises a special value 'auto'. If given, the Mix task will try to
automatically deduce the MinIO build matching the current machine.

Together with '--version latest' this permits running the
minio_server.download task without requiring user intervention.

* Mention '--arch auto' in README